refactor(saga-output): set up saga output for awaitable
